Module move_bytecode_utils::layout

Structs

SerdeLayoutBuilder

Type for building a registry of serde-reflection friendly struct layouts for Move types. The layouts created by this type are intended to be passed to the serde-generate tool to create struct bindings for Move types in source languages that use Move-based services. The LayoutBuilder can operate in two modes: “deep” and “shallow”. In shallow mode, it will generate a single layout for the struct or type passed in by the user (under the assumption that layouts for dependencies have been generated previously). In deep mode, it will generate layouts for all of the (transitive) dependencies of the type passed in, as well as layouts for the Move ground types like address and signer. The result is a self-contained registry with no unbound typenames

Enums

StructLayoutBuilder

TypeLayoutBuilder