Trait move_bytecode_utils::module_cache::GetModule

pub trait GetModule {
    type Error: Debug;
    type Item: Borrow<CompiledModule>;

    fn get_module_by_id(
        &self, 
        id: &ModuleId
    ) -> Result<Option<Self::Item>, Self::Error>;
}

A persistent storage that can fetch the bytecode for a given module id TODO: do we want to implement this in a way that allows clients to cache struct layouts?

Required Associated Types

type Error: Debug

type Item: Borrow<CompiledModule>

Required Methods

fn get_module_by_id(
&self,
id: &ModuleId
) -> Result<Option<Self::Item>, Self::Error>